Some progress

+++ usac_tests.py
@@ -25,25 +25,29 @@ ji_env = ji.Environment(
     loader=ji.FileSystemLoader(os.path.join(os.path.dirname(os.path.realpath(__file__)), 'tex_templates')))
 
 
-def compile_tex(rendered_tex, out_pdf_path):
-    tmp_dir = tempfile.mkdtemp()
-    in_tmp_path = os.path.join(tmp_dir, 'rendered.tex')
-    with open(in_tmp_path, 'w') as outfile:
+def compile_tex(rendered_tex, out_tex_dir, out_tex_file, out_pdf_filen=None):
+    texdf = os.path.join(out_tex_dir, out_tex_file)
+    with open(texdf, 'w') as outfile:
         outfile.write(rendered_tex)
-    out_tmp_path = os.path.join(tmp_dir, 'out.pdf')
-    cmdline = ['pdflatex', in_tmp_path, '-job-name', 'out', '-output-directory', tmp_dir]
-    try:
-        retcode = sp.run(cmdline, shell=False, check=True).returncode
-        if retcode < 0:
-            print("Child pdflatex was terminated by signal", -retcode, file=sys.stderr)
-        else:
-            print("Child returned", retcode)
-    except OSError as e:
-        print("Execution of pdflatex failed:", e, file=sys.stderr)
-    except sp.CalledProcessError as e:
-        print("Execution of pdflatex failed:", e, file=sys.stderr)
-    shutil.copy(out_tmp_path, out_pdf_path)
-    shutil.rmtree(tmp_dir)
+    if out_pdf_filen is not None:
+        tmp_dir = tempfile.mkdtemp()
+        shutil.copy(texdf, tmp_dir)
+        tex_tmp = os.path.join(tmp_dir, out_tex_file)
+        pdfpath, pdfname = os.path.split(out_pdf_filen)
+        out_tmp_path = os.path.join(tmp_dir, pdfname)
+        cmdline = ['pdflatex', tex_tmp, '-job-name', pdfname, '-output-directory', tmp_dir]
+        try:
+            retcode = sp.run(cmdline, shell=False, check=True).returncode
+            if retcode < 0:
+                print("Child pdflatex was terminated by signal", -retcode, file=sys.stderr)
+            else:
+                print("Child returned", retcode)
+        except OSError as e:
+            print("Execution of pdflatex failed:", e, file=sys.stderr)
+        except sp.CalledProcessError as e:
+            print("Execution of pdflatex failed:", e, file=sys.stderr)
+        shutil.copy(out_tmp_path, pdfpath)
+        shutil.rmtree(tmp_dir)
 
 
 def calcSatisticRt_th(data, store_path):
@@ -60,6 +64,11 @@ def calcSatisticRt_th(data, store_path):
     if data.empty:
         raise ValueError('No data left after filtering unsuccessful estimations')
 
+    pdf_folder = os.path.join(store_path, 'pdf')
+    try:
+        os.mkdir(pdf_folder)
+    except FileExistsError:
+        print('Folder', pdf_folder, 'for storing pdf files already exists')
     tex_folder = os.path.join(store_path, 'tex')
     try:
         os.mkdir(tex_folder)
@@ -79,6 +88,7 @@ def calcSatisticRt_th(data, store_path):
     errvalnames = stats.columns.values # Includes statistic name and error value names
     grp_names = stats.index.names #As used when generating the groups
     rel_data_path = os.path.relpath(tex_folder, tdata_folder)
+    texf_name = 'data_USAC_opts_' + grp_names[0] + '-' + grp_names[1] + '_combs_vs_' + grp_names[2] + '.tex'
     # holds the grouped names/entries within the group names excluding the last entry th
     #grp_values = list(dict.fromkeys([i[0:2] for i in stats.index.values]))
     tex_infos = {'title': 'Statistics for USAC Option Combinations of ' + grp_names[0] + ' and ' +
@@ -97,10 +107,10 @@ def calcSatisticRt_th(data, store_path):
             ## tmp.index = ['%s' % str(c) for a, b, c in tmp.index]
             ## tmp.index = ['%s' % str(b) for a, b in tmp.index]
             ## tmp.index.name = str(grp_names[-1])
-            tex_name = 'tex_' + '_'.join(map(str, it)) + '_vs_' + \
+            dataf_name = 'data_' + '_'.join(map(str, it)) + '_vs_' + \
                        str(grp_names[-1]) + '.csv'
-            tex_name = tex_name.replace('%', 'perc')
-            ftex_name = os.path.join(tdata_folder, tex_name)
+            dataf_name = dataf_name.replace('%', 'perc')
+            ftex_name = os.path.join(tdata_folder, dataf_name)
             # with open(ftex_name, 'a') as f:
             #     f.write('# ' + str(it[-1]) + ' values for ' + str(it[0]) + '\n')
             #     f.write('# Column parameters: ' + '-'.join(grp_names[0:2]) + '\n')
@@ -119,17 +129,21 @@ def calcSatisticRt_th(data, store_path):
                stats_all['max'][0] > (stats_all['mean'][0] + stats_all['std'][0] * 2.576):
                 use_limits = [round(stats_all['mean'][0] - stats_all['std'][0] * 2.576, 6),
                               round(stats_all['mean'][0] + stats_all['std'][0] * 2.576, 6)]
-            reltex_name = os.path.join(rel_data_path, tex_name)
+            reltex_name = os.path.join(rel_data_path, dataf_name)
             tex_infos['sections'].append({'file': reltex_name,
                                           'name': replace_stat_names(it[-1]) + ' values for ' + str(it[0]) +
                                                   ' compared to ' + str(grp_names[-1]),
                                           'fig_type': 'smooth',
-                                          'plots_y': list(tmp.columns.values),
+                                          'plots': list(tmp.columns.values),
+                                          'axis_y': replace_stat_names(it[-1]),
                                           'plot_x': str(grp_names[-1]),
                                           'limits': use_limits
                                           })
-            template = ji_env.get_template('usac-testing_single_plots.tex')
-            #tmp = tempfile.mkdtemp()
+
+    template = ji_env.get_template('usac-testing_single_plots.tex')
+    rendered_tex = template.render(title = tex_infos['title'],
+                                   sections = tex_infos['sections'])
+    compile_tex(rendered_tex, tex_folder, texf_name)
 
 
 def replace_stat_names(name):
